requirements.txt
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.1
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.9.1
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
six==1.16.0
sklearn-model==0.0.6
threadpoolctl==3.5.0
tzdata==2024.1
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.1
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.9.1
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
six==1.16.0
sklearn-model==0.0.6
threadpoolctl==3.5.0
tzdata==2024.1
